+++ src/Notifier.php
@@ -58,6 +58,13 @@ class Notifier
             throw new InvalidNotificationException($notification, 'Notification body can not be empty');
         }
 
+        // This makes the icon accessible for native commands when it's embedded inside a phar
+        if (($icon = $notification->getIcon()) && 0 === strpos($icon, 'phar://')) {
+            $notification->setIcon(
+                $this->extractIconFromPhar($icon)
+            );
+        }
+
         return $this->driverInUse->send($notification);
     }
 
@@ -87,5 +94,24 @@ class Notifier
         }
 
         return $bestDriver;
+    }
+
+    /**
+     * Extract the icon from the phar archive to make it accessible for native commands.
+     *
+     * @param string $icon
+     *
+     * @return string
+     */
+    private function extractIconFromPhar($icon)
+    {
+        $pharPath         = \Phar::running(false);
+        $iconRelativePath = substr($icon, strpos($icon, $pharPath) + strlen($pharPath) + 1);
+        $tmpDir           = sys_get_temp_dir().'/jolinotif';
+
+        $phar = new \Phar($pharPath);
+        $phar->extractTo($tmpDir, $iconRelativePath, true);
+
+        return $tmpDir.'/'.$iconRelativePath;
     }
 }