+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 * added NotifySendDriver (can be used on most Linux distributions)
 * added TerminalNotifierDriver (can be used on Mac OS X 10.8 and higher)
 * added GrowlNotifyDriver (can be used on Mac OS if growlnotify is available)
+* added icon extract when provided from a phar archive